requirements.txt
dotenv~=0.9.9
python-dotenv~=1.2.1
pydantic~=2.12.5
tomli~=2.3.0
pytest~=9.0.1
uvicorn~=0.38.0
fastapi~=0.123.5
google-genai
gunicorn
pydantic-settings
sqlalchemy